+++ tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "baseUrl": "./src",
     "outDir": "dist",
     "module": "esnext",
     "lib": ["dom", "esnext"],
